+++ tests/fixtures/application_fixtures.py
@@ -2831,8 +2831,8 @@ def tcf_system(db: Session) -> System:
 
 
 @pytest.fixture(scope="function")
-def ac_system(db: Session) -> System:
-    """Test AC System - will be fleshed out further later"""
+def ac_system_with_privacy_declaration(db: Session) -> System:
+    """Test AC System with a privacy declaration"""
     system = System.create(
         db=db,
         data={
@@ -2856,6 +2856,23 @@ def ac_system(db: Session) -> System:
             ],
         },
     )
+    return system
+
+
+@pytest.fixture(scope="function")
+def ac_system_without_privacy_declaration(db: Session) -> System:
+    """Test AC System without privacy declaration"""
+    system = System.create(
+        db=db,
+        data={
+            "fides_key": f"ac_system{uuid.uuid4()}",
+            "vendor_id": "gacp.100",
+            "name": f"Test AC System",
+            "organization_fides_key": "default_organization",
+            "system_type": "Service",
+        },
+    )
+
     return system